# Conflicts:
#	scripts/README.md
#	scripts/build_apk.bat
#	scripts/build_apk.sh
#	scripts/build_linux.sh
#	scripts/build_macos.sh
#	scripts/build_web.bat
#	scripts/build_web.sh
#	scripts/build_windows.bat
#	scripts/build_windows.sh

+++ business/system/lib/system/repository/remote/sys_oss_api.dart
@@ -47,7 +47,7 @@ class SysOssRepository {
 
   ///删除OSS存储
   static Future<IntensifyEntity<dynamic>> delete(CancelToken cancelToken,
-      {int? id, List<int>? ids}) {
+      {BigInt? id, List<BigInt>? ids}) {
     //参数校验
     assert(id != null && ids == null || id == null && ids != null);
     ids ??= [id!];
