+++ business/system/lib/system/ui/dict/type/dict_type_list_browser_page.dart
@@ -171,7 +171,7 @@ class _DictTypeListBrowserVm extends AppBaseVm {
   }
 
   //删除事件
-  void onDelete({Future<bool?> Function(List<String>)? confirmHandler, List<int>? idList}) {
+  void onDelete({Future<bool?> Function(List<String>)? confirmHandler, List<BigInt>? idList}) {
     if (idList == null) {
       List<SysDictType> selectList = SelectUtils.getSelect(listVmSub.dataList) ?? [];
       if (selectList.isEmpty) {
@@ -179,7 +179,7 @@ class _DictTypeListBrowserVm extends AppBaseVm {
         return;
       }
       List<String> nameList = selectList.map<String>((item) => item.dictName!).toList();
-      List<int> idList = selectList.map<int>((item) => item.dictId!).toList();
+      List<BigInt> idList = selectList.map<BigInt>((item) => item.dictId!).toList();
       confirmHandler?.call(nameList).then((value) {
         if (value == true) {
           onDelete(idList: idList);
