+++ business/system/lib/system/ui/notice/notice_list_browser_page.dart
@@ -166,7 +166,7 @@ class _NoticeListBrowserVm extends AppBaseVm {
   }
 
   //删除事件
-  void onDelete({Future<bool?> Function(List<String>)? confirmHandler, List<int>? idList}) {
+  void onDelete({Future<bool?> Function(List<String>)? confirmHandler, List<BigInt>? idList}) {
     if (idList == null) {
       List<SysNotice> selectList = SelectUtils.getSelect(listVmSub.dataList) ?? [];
       if (selectList.isEmpty) {
@@ -174,7 +174,7 @@ class _NoticeListBrowserVm extends AppBaseVm {
         return;
       }
       List<String> nameList = selectList.map<String>((item) => item.noticeTitle!).toList();
-      List<int> idList = selectList.map<int>((item) => item.noticeId!).toList();
+      List<BigInt> idList = selectList.map<BigInt>((item) => item.noticeId!).toList();
       confirmHandler?.call(nameList).then((value) {
         if (value == true) {
           onDelete(idList: idList);
