+++ business/system/lib/system/ui/oss/config/oss_config_list_browser_page.dart
@@ -159,7 +159,7 @@ class _OssConfigListBrowserVm extends AppBaseVm {
   }
 
   //删除事件
-  void onDelete({Future<bool?> Function(List<String>)? confirmHandler, List<int>? idList}) {
+  void onDelete({Future<bool?> Function(List<String>)? confirmHandler, List<BigInt>? idList}) {
     if (idList == null) {
       List<SysOssConfig> selectList = SelectUtils.getSelect(listVmSub.dataList) ?? [];
       if (selectList.isEmpty) {
@@ -167,7 +167,7 @@ class _OssConfigListBrowserVm extends AppBaseVm {
         return;
       }
       List<String> nameList = selectList.map<String>((item) => item.configKey!).toList();
-      List<int> idList = selectList.map<int>((item) => item.ossConfigId!).toList();
+      List<BigInt> idList = selectList.map<BigInt>((item) => item.ossConfigId!).toList();
       confirmHandler?.call(nameList).then((value) {
         if (value == true) {
           onDelete(idList: idList);
