+++ business/system/lib/system/ui/tenant/tenant_list_browser_page.dart
@@ -202,7 +202,7 @@ class _TenantListBrowserVm extends AppBaseVm with CancelTokenAssist {
   }
 
   //删除事件
-  void onDelete({Future<bool?> Function(List<String>)? confirmHandler, List<int>? idList}) {
+  void onDelete({Future<bool?> Function(List<String>)? confirmHandler, List<BigInt>? idList}) {
     if (idList == null) {
       List<SysTenant> selectList = SelectUtils.getSelect(listVmSub.dataList) ?? [];
       if (selectList.isEmpty) {
@@ -210,7 +210,7 @@ class _TenantListBrowserVm extends AppBaseVm with CancelTokenAssist {
         return;
       }
       List<String> nameList = selectList.map<String>((item) => item.companyName!).toList();
-      List<int> idList = selectList.map<int>((item) => item.id!).toList();
+      List<BigInt> idList = selectList.map<BigInt>((item) => item.id!).toList();
       confirmHandler?.call(nameList).then((value) {
         if (value == true) {
           onDelete(idList: idList);
